[STC-1] Added password reset views and controllers.
