+++ scriptoocloud/src/main/java/edu/asu/diging/scriptoocloud/web/manager/AdminController.java
@@ -1,0 +1,10 @@
+package edu.asu.diging.scriptoocloud.web.manager;
+
+import org.springframework.stereotype.Controller;
+
+@Controller
+public class AdminController{
+
+  
+    
+}
