+++ scriptoocloud/src/main/webapp/WEB-INF/views/reset/sent.html
@@ -4,7 +4,7 @@
   </head>
   <body>
   <div layout:fragment="content">
-  <h2>Sent!</h2>
+  <h2>Password Reset Sent to email!</h2>
   
 
 </div>
